Restructure client integration tests
